PROJETOS/Projeto Subs Generator with AI/requirements.txt
openai-whisper==20240930
torch==1.10.1
deep-translator==1.11.4